requirements.txt
ttkbootstrap
reportlab
pyhanko
python-jose